induction/requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
Deprecated==1.2.13
Django==3.2.16
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==21.1
django-redis==5.2.0
django-rest-framework==0.1.0
django-seed==0.3.1
django-session-timeout==0.1.0
django-six==1.0.4
djangorestframework==3.14.0
Faker==14.2.1
hiredis==2.0.0
importlib-metadata==4.8.3
Markdown==3.3.7
model-bakery==1.9.0
packaging==21.3
Pillow==8.4.0
pkg_resources==0.0.0
psycopg2-binary==2.9.4
pymemcache==3.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
redis==4.3.4
redis-simple-cache==0.0.8
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
toposort==1.7
typing_extensions==4.1.1
wrapt==1.14.1
xlwt==1.3.0
zipp==3.6.0
